[aml] rename Amlogic.xxx to AMLCodec.xxx
